pyproject.toml
[project]
name = "get-searched-reply"
version = "0.1.0"
description = "A CLI skill for agents to get AI-generated search summaries using Alibaba Cloud's model-side search."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.33.0",
    "pyyaml>=6.0.3",
    "typer>=0.25.0",
]

[project.scripts]
search = "main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
